Creating a small example to show an event driven architecture.
The events for the user are persisted via the event-store and the listener gets a notification after a new event was sent.

Based on the event the present state of the user will be stored in my simple in memory user-store. The present state can be read by using the UserReadRepository.

The test-class will show the full flow.